requirements.txt
pyvisa
pandas
pyserial
python-vxi11